Refactor login 유지 상태 분리
