+++ components/core/layout/Layout.tsx
@@ -1,13 +1,9 @@
 import styled from "styled-components";
-import { ReactNode, useEffect } from "react";
+import {ReactNode} from "react";
 import NavigationHeader from "../../containers/navigation-header/NavigationHeader";
 import { useAxiosInterceptor } from "../../../utils/hooks/useAxiosInterceptor";
-import { useDispatch, useSelector } from "react-redux";
 import ModalManager from "../../modal/modal/Modal";
-import { loginMember } from "../../../api/member";
-import { useCookies } from "react-cookie";
-import { useRouter } from "next/router";
-import { loginUser, selectUser } from "../../../store/slices/user/user";
+import useLoginEffect from "../../../utils/useLoginEffect";
 
 interface LayoutProps {
   children: ReactNode;
@@ -20,43 +16,7 @@ export const Container = styled.div`
 
 function Layout({ children }: LayoutProps) {
   useAxiosInterceptor();
-  const router = useRouter();
-  const { query } = router;
-  const [cookies, setCookies, removeCookie] = useCookies(["token"]);
-  const dispatch = useDispatch();
-  const { isLogin } = useSelector(selectUser);
-
-  useEffect(() => {
-    const { token } = query;
-    if (token) {
-      setCookies("token", token, { path: "/", httpOnly: true });
-    }
-
-    const jwt = token || cookies["token"];
-
-    if (isLogin || !jwt) return;
-    setCookies("token", jwt, { path: "/" });
-    const getUser = async () => {
-      try {
-        const { data } = await loginMember();
-        dispatch(loginUser(data));
-      } catch (err) {
-        removeCookie("token");
-      } finally {
-        router.push("/");
-      }
-    };
-    getUser();
-  }, [
-    isLogin,
-    dispatch,
-    setCookies,
-    query,
-    router,
-    cookies,
-    cookies.token,
-    removeCookie,
-  ]);
+  useLoginEffect();
 
   return (
     <Container>
